requirements.txt
requests==2.27.1
firebase-admin==5.2.0
python-dateutil==2.8.2
flask[async]
structlog==21.5.0
flask-cors
aiodns==3.0.0
aiohttp==3.7.4.post0
ed25519==1.5
base58==2.1.1
google-cloud-tasks==2.9.0
google-cloud-storage==2.4.0
near-api-py==0.1.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
